Remove make check from the .spec
Makes the build take ages and at the moment the library is used only
by tracker-miners so maybe not critical. More commonly the the checks
are in unit test package and executed on the device.

Though locally having one failure at the moment:
--
Test 52-basic-load%%001-00001 result:   FAILURE 52-basic-load rc=125
 batch name: 53-sim-binary_tree
 test mode:  c
 test type:  bpf-sim
 test arch:  x86_64
--

Is that related to the build environment, no idea. Anyhow we don't build
x86_64 currently.